refactor(pdf-parser): extract and modularize text-reference matching logic
Move `matchTextToReference` and associated utilities to a new `text-reference-matcher` module for better separation of concerns and code reuse. Simplify `vlm-text-corrector` by delegating matching operations to the new utility. Add comprehensive unit tests for the extracted functions and refactor related test files for clarity and coverage.
