+++ packages/pdf-parser/src/processors/vlm-text-corrector.test.ts
@@ -1640,199 +1640,6 @@ describe('VlmTextCorrector', () => {
     });
   });
 
-  describe('matchTextToReference', () => {
-    test('matches garbled footnote to correct pdftotext line', () => {
-      const pageTexts = [
-        {
-          index: 0,
-          item: createTextItem(
-            '49) (W)#X1CR003T 2008, 『아산 상성리유적』.',
-            'footnote',
-            1,
-          ),
-        },
-      ];
-      const pageText = '49) (財)忠淸文化財硏究院 2008,『아산 상성리유적』.';
-
-      const result = corrector.matchTextToReference(pageTexts, pageText);
-
-      expect(result.size).toBe(1);
-      expect(result.get(0)).toBe(
-        '49) (財)忠淸文化財硏究院 2008,『아산 상성리유적』.',
-      );
-    });
-
-    test('matches heavily garbled text above threshold', () => {
-      const pageTexts = [
-        {
-          index: 0,
-          item: createTextItem(
-            '50) (M):23x1CR03% 2008, 『아산 장재리 아골유적』.',
-            'footnote',
-            1,
-          ),
-        },
-      ];
-      const pageText =
-        '50) (財)忠淸文化財硏究院 2008,『아산 장재리 아골유적』.';
-
-      const result = corrector.matchTextToReference(pageTexts, pageText);
-
-      expect(result.size).toBe(1);
-      expect(result.get(0)).toBe(pageText);
-    });
-
-    test('skips identical text (no ref needed)', () => {
-      const pageTexts = [
-        {
-          index: 0,
-          item: createTextItem('제1장 조사개요', 'section_header', 1),
-        },
-      ];
-      const pageText = '제1장 조사개요';
-
-      const result = corrector.matchTextToReference(pageTexts, pageText);
-
-      expect(result.size).toBe(0);
-    });
-
-    test('returns empty map when no matches above threshold', () => {
-      const pageTexts = [
-        {
-          index: 0,
-          item: createTextItem('completely different text', 'text', 1),
-        },
-      ];
-      const pageText = 'XXXXXXXXYYYYYYZZZZZZ';
-
-      const result = corrector.matchTextToReference(pageTexts, pageText);
-
-      expect(result.size).toBe(0);
-    });
-
-    test('returns empty map for empty pageText', () => {
-      const pageTexts = [
-        { index: 0, item: createTextItem('some text', 'text', 1) },
-      ];
-
-      const result = corrector.matchTextToReference(pageTexts, '');
-
-      expect(result.size).toBe(0);
-    });
-
-    test('handles greedy matching without double assignment', () => {
-      const pageTexts = [
-        { index: 0, item: createTextItem('AAA BBB', 'text', 1) },
-        { index: 1, item: createTextItem('CCC DDD', 'text', 1) },
-        { index: 2, item: createTextItem('EEE FFF', 'text', 1) },
-      ];
-      const pageText = 'AAA BBB ref\n\nCCC DDD ref\n\nEEE FFF ref';
-
-      const result = corrector.matchTextToReference(pageTexts, pageText);
-
-      expect(result.size).toBe(3);
-      // Each element should match a unique ref block
-      const refValues = new Set(result.values());
-      expect(refValues.size).toBe(3);
-    });
-
-    test('handles empty OCR text element against non-empty ref line', () => {
-      const pageTexts = [{ index: 0, item: createTextItem('', 'text', 1) }];
-      const pageText = 'some reference text';
-
-      const result = corrector.matchTextToReference(pageTexts, pageText);
-
-      expect(result.size).toBe(0);
-    });
-
-    test('handles more OCR elements than reference lines', () => {
-      const pageTexts = [
-        { index: 0, item: createTextItem('AAA BBB', 'text', 1) },
-        { index: 1, item: createTextItem('CCC DDD', 'text', 1) },
-        { index: 2, item: createTextItem('EEE FFF', 'text', 1) },
-      ];
-      const pageText = 'AAA BBB ref';
-
-      const result = corrector.matchTextToReference(pageTexts, pageText);
-
-      // Only one ref block available, so at most 1 match
-      expect(result.size).toBeLessThanOrEqual(1);
-    });
-
-    test('matches long OCR paragraph against merged pdftotext block', () => {
-      // OCR produces one long paragraph; pdftotext splits into layout lines
-      const ocrText =
-        '唐은 熊津(공주), 馬韓(익산), 東明에 都督府를 설치하고 9州 5小京制를 완성하였다';
-      const pageTexts = [
-        { index: 0, item: createTextItem(ocrText, 'text', 1) },
-      ];
-      // pdftotext: same content split across multiple lines (no blank line separator)
-      const pageText =
-        '唐은 熊津(공주), 馬韓(익산),\n東明에 都督府를 설치하고\n9州 5小京制를 완성하였다';
-
-      // Identical after merge → no ref needed (skips identical text)
-      const identicalResult = corrector.matchTextToReference(
-        pageTexts,
-        pageText,
-      );
-      expect(identicalResult.size).toBe(0);
-
-      // Use a garbled OCR version to verify ref is provided
-      const garbledOcrText =
-        '받은 M(공주), 5류(익산), 햇배에 Bbt를 설치하고 9MM 5☆를 완성하였다';
-      const garbledPageTexts = [
-        { index: 0, item: createTextItem(garbledOcrText, 'text', 1) },
-      ];
-
-      const garbledResult = corrector.matchTextToReference(
-        garbledPageTexts,
-        pageText,
-      );
-
-      expect(garbledResult.size).toBe(1);
-      expect(garbledResult.get(0)).toBe(
-        '唐은 熊津(공주), 馬韓(익산), 東明에 都督府를 설치하고 9州 5小京制를 완성하였다',
-      );
-    });
-
-    test('separates blocks at blank lines for independent matching', () => {
-      const pageTexts = [
-        {
-          index: 0,
-          item: createTextItem(
-            '본문 단락 텍스트 내용이 여기에 있습니다',
-            'text',
-            1,
-          ),
-        },
-        {
-          index: 1,
-          item: createTextItem('49) (W)#X1 2008, 『보고서』.', 'footnote', 1),
-        },
-      ];
-      // pdftotext: body paragraph (2 layout lines) + blank line + footnote
-      const pageText =
-        '본문 단락 텍스트\n내용이 여기에 있습니다\n\n49) (財)忠淸 2008, 『보고서』.';
-
-      const result = corrector.matchTextToReference(pageTexts, pageText);
-
-      // Body paragraph matches merged block, footnote matches separate block
-      expect(result.size).toBe(1);
-      expect(result.get(1)).toBe('49) (財)忠淸 2008, 『보고서』.');
-    });
-
-    test('returns empty map when pdftotext contains only blank lines', () => {
-      const pageTexts = [
-        { index: 0, item: createTextItem('some text', 'text', 1) },
-      ];
-      const pageText = '\n\n\n\n';
-
-      const result = corrector.matchTextToReference(pageTexts, pageText);
-
-      expect(result.size).toBe(0);
-    });
-  });
-
   describe('system prompt content', () => {
     function setupProcessFn() {
       vi.mocked(ConcurrentPool.run).mockImplementation(
